fix(synthetic-shadow): ekashida feedback
