+++ packages/@lwc/engine/src/framework/hooks.ts
@@ -16,7 +16,7 @@ import {
     appendVM,
     runWithBoundaryProtection,
 } from './vm';
-import { VNode, VCustomElement, VElement } from '../3rdparty/snabbdom/types';
+import { VNode, VCustomElement, VElement, VNodes } from '../3rdparty/snabbdom/types';
 import modEvents from './modules/events';
 import modAttrs from './modules/attrs';
 import modProps from './modules/props';
@@ -25,7 +25,6 @@ import modComputedStyle from './modules/computed-style-attr';
 import modStaticClassName from './modules/static-class-attr';
 import modStaticStyle from './modules/static-style-attr';
 import modContext from './modules/context';
-import { hasDynamicChildren } from './patch';
 import { updateDynamicChildren, updateStaticChildren } from '../3rdparty/snabbdom/snabbdom';
 import {
     patchCustomElementWithRestrictions,
@@ -271,4 +270,17 @@ export function removeCustomElmHook(vnode: VCustomElement) {
     // for custom elements we don't have to go recursively because the removeVM routine
     // will take care of disconnecting any child VM attached to its shadow as well.
     removeVM(getCustomElementVM(vnode.elm as HTMLElement));
+}
+
+// Using a WeakMap instead of a WeakSet because this one works in IE11 :(
+const FromIteration: WeakMap<VNodes, 1> = new WeakMap();
+
+// dynamic children means it was generated by an iteration
+// in a template, and will require a more complex diffing algo.
+export function markAsDynamicChildren(children: VNodes) {
+    FromIteration.set(children, 1);
+}
+
+export function hasDynamicChildren(children: VNodes): boolean {
+    return FromIteration.has(children);
 }
