--- packages/@lwc/engine/src/framework/patch.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-import { VNodes } from '../3rdparty/snabbdom/types';
-
-// Using a WeakMap instead of a WeakSet because this one works in IE11 :(
-const FromIteration: WeakMap<VNodes, 1> = new WeakMap();
-
-// dynamic children means it was generated by an iteration
-// in a template, and will require a more complex diffing algo.
-export function markAsDynamicChildren(children: VNodes) {
-    FromIteration.set(children, 1);
-}
-
-export function hasDynamicChildren(children: VNodes): boolean {
-    return FromIteration.has(children);
-}
