+++ packages/@lwc/synthetic-shadow/README.md
@@ -7,6 +7,8 @@ This is a polyfill for ShadowRoot that was tailor-made for LWC in order to meet 
 ## Compromises
 
 -   Default content for `<slot>` elements is always empty.
+-   `slotchange` is only available directly on the slot (it doesn't bubble as in the case of the native implementation)
+-   Dispatch events directly on the ShadowRoot instance are not allowed because they will always leak into the host, and we have no way to contain them.
 
 ## Missing features
